[CSL-909] Move useful functions out of WALLET_MODE
We used to have MonadBalances, MonadTxHistory, sendTx, etc. only available in
WALLET_MODE, even though they didn't depend on wallet and they were useful
outside of WALLET_MODE (e.g. for the smartgen). This commit moves them from
Pos.Wallet.WalletMode into the following modules:

* Pos.Balances         – `MonadBalances`
* Pos.TxHistory        – `MonadTxHistory`
* Pos.Util.Tx          – transaction construction (e.g. `makePubKeyTx`)
* Pos.Communication.Tx – `TxMode` and transaction sending (e.g. `submitTx`)
