+++ cardano-sl.cabal
@@ -71,6 +71,8 @@ library
                         Pos.Security.CLI
                         Pos.Security.Util
                         Pos.Security.Workers
+                        Pos.TxHistory
+                        Pos.Balances
 
                         -- SSC
                         Pos.Ssc.Arbitrary
@@ -103,6 +105,7 @@ library
                         Pos.Util.BackupPhrase
                         Pos.Util.JsonLog
                         Pos.Util.UserSecret
+                        Pos.Util.Tx
 
                         -- Worker, work mode
                         Pos.Worker
@@ -181,6 +184,7 @@ library
                         Pos.Communication.Types
                         Pos.Communication.Message
                         Pos.Communication.Methods
+                        Pos.Communication.Tx
                         Pos.Communication.Server
                         Pos.Communication.Server.SysStart
                         Pos.Communication.Specs
@@ -352,9 +356,7 @@ library
                         Pos.Web.Server
                         Pos.Web.Types
   if flag(with-wallet)
-    other-modules:      Pos.Wallet.Tx
-                        Pos.Wallet.Tx.Pure
-                        Pos.Wallet.WalletMode
+    other-modules:      Pos.Wallet.WalletMode
                         Pos.Wallet.Launcher
                         Pos.Wallet.Launcher.Runner
                         Pos.Wallet.Launcher.Param
