+++ src/Pos/Communication.hs
@@ -12,5 +12,6 @@ import           Pos.Communication.Protocol  as M
 import           Pos.Communication.Relay     as M
 import           Pos.Communication.Server    as M
 import           Pos.Communication.Specs     as M
+import           Pos.Communication.Tx        as M
 import           Pos.Communication.Types     as M
 import           Pos.Communication.Util      as M
