+++ src/Pos/Wallet.hs
@@ -1,11 +1,9 @@
 module Pos.Wallet
        ( module Pos.Wallet.Launcher
-       , module Pos.Wallet.Tx
        , module Pos.Wallet.Update
        , module Pos.Wallet.WalletMode
        ) where
 
 import           Pos.Wallet.Launcher
-import           Pos.Wallet.Tx
 import           Pos.Wallet.Update
 import           Pos.Wallet.WalletMode
