+++ src/wallet/Main.hs
@@ -26,7 +26,7 @@ import           Universum
 import           Pos.Binary                (Raw)
 import qualified Pos.CLI                   as CLI
 import           Pos.Communication         (OutSpecs, SendActions, Worker', WorkerSpec,
-                                            worker)
+                                            sendTxOuts, submitTx, worker)
 import           Pos.Crypto                (Hash, SecretKey, createProxySecretKey,
                                             fakeSigner, hash, hashHexF, sign, toPublic,
                                             unsafeHash)
@@ -48,8 +48,8 @@ import           Pos.Update                (BlockVersionData (..), UpdateProposa
                                             skovorodaUpdateData)
 import           Pos.Wallet                (WalletMode, WalletParams (..), WalletRealMode,
                                             getBalance, runWalletReal, sendProposalOuts,
-                                            sendTxOuts, sendVoteOuts, submitTx,
-                                            submitUpdateProposal, submitVote)
+                                            sendVoteOuts, submitUpdateProposal,
+                                            submitVote)
 #ifdef WITH_WEB
 import           Pos.Wallet.Web            (walletServeWebLite, walletServerOuts)
 #endif
